[修改] 1、对用户还书，通知关注这本书的用户去借书的邮件提示进行修改，增加了图书名字的显示！

+++ BookManager/src/main/java/com/bookmanager/book/controller/BookAdminController.java
@@ -62,7 +62,7 @@ public class BookAdminController {
      */
     @UserLoginToken
     @PostMapping("/delete/{isbn}")
-    public Result deleteBook(@PathVariable Long isbn){
+    public Result deleteBook(@PathVariable Integer isbn){
         return bookService.deleteBookByIsbn(isbn);
     }
